requirements.txt
scikit-learn==1.0.2
pickleshare==0.7.5
ninja==1.10.2.3
gdown
easydict==1.9
PyYAML==6.0
protobuf==3.19.4
tensorboard==2.8.0
termcolor==1.1.0
tqdm==4.62.3
multimethod==1.7
h5py==3.6.0
matplotlib==3.5.1
wandb
pyvista
setuptools==59.5.0
Cython==0.29.28
pandas
deepspeed
shortuuid
laspy               # For reading and writing .laz files of point cloud data
laszip              # For compressing and decompressing .laz files

# for docs
mkdocs-material
mkdocs-awesome-pages-plugin
mdx_truly_sane_lists
